Add unit tests for RowObjectHelpers and enhance field enable/disable logic
- Created a comprehensive test suite for RowObjectHelpers, covering methods related to row and field manipulation.
- Enhanced the logic in SetDisabledField and SetEnabledField methods to check if the field state has changed before modifying the RowAction.
- Updated DisableAllFieldObjects and EnableAllFieldObjects methods to only set the RowAction if changes occur.
